api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Thanh Nguyen Hoang",email = "thanh.nguyenhoang@paradox.ai"}
]
readme = "./README.md"
requires-python = ">=3.11"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.11"
uvicorn = "^0.23.0"
fastapi = "^0.110.0"  
requests = "^2.31.0" 
python-multipart = "^0.0.7"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
websockets = "^10.0" 